+++ CHANGELOG.md
@@ -17,3 +17,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - `#count` method.
     - `#my_map` method.
     - `#my_map!` method.
+    - `#my_inject` method.